+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - #137 Adds key binding editor
 - #151 Moves next undo/redo actions to separate bar, adds setting to toggle showing it
 - Adds Open Koffee Data Folder command
+- Adds centered message display for empty folders and search results instead of "empty" item
 - Adds mouse tooltips to status, yank register, next undo/redo action
 - Improves resiliency of loading and saving config and history
 - Changes recycle/delete to not push useless undo actions